[REFACTORING]: moved VitalityLoader to shared-ui
