+++ v6y-apps/front-bo/src/infrastructure/components/RefineTableWrapper.tsx
@@ -1,8 +1,8 @@
 import { List, RefreshButton, useTable } from '@refinedev/antd';
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import { VitalityTitle } from '@v6y/shared-ui';
 import { ReactNode } from 'react';
 
-import VitalityLoader from '../../commons/components/VitalityLoader';
 import { RefineTableType } from '../types/RefineTableType';
 
 export default function RefineTableWrapper<T>({
