+++ v6y-apps/front/src/app/layout.tsx
@@ -1,7 +1,7 @@
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import * as React from 'react';
 import { ReactNode, Suspense } from 'react';
 
-import VitalityLoader from '../commons/components/VitalityLoader';
 import VitalityPageLayout from '../commons/components/layout/VitalityPageLayout';
 import VitalityTerms from '../commons/config/VitalityTerms';
 import VitalityTheme from '../commons/config/VitalityTheme';
