+++ v6y-apps/front/src/commons/components/VitalityDynamicLoader.tsx
@@ -1,7 +1,6 @@
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import dynamic, { DynamicOptions } from 'next/dynamic';
 import * as React from 'react';
-
-import VitalityLoader from './VitalityLoader';
 
 interface DynamicComponentProps {
     [key: string]: unknown;
