+++ v6y-apps/front/src/commons/components/VitalitySectionView.tsx
@@ -1,4 +1,5 @@
 import { ExportOutlined } from '@ant-design/icons';
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import { VitalityTitle } from '@v6y/shared-ui';
 import { Avatar, Button, Card, Col, Row, Space, Typography } from 'antd';
 import * as React from 'react';
@@ -6,7 +7,6 @@ import * as React from 'react';
 import vitalityTheme from '../config/VitalityTheme';
 import { VitalitySectionViewProps } from '../types/VitalitySectionViewProps';
 import VitalityEmptyView from './VitalityEmptyView';
-import VitalityLoader from './VitalityLoader';
 
 const VitalitySectionView = ({
     isLoading,
