+++ v6y-apps/front/src/commons/components/indicators/VitalitySelectableIndicators.tsx
@@ -1,4 +1,5 @@
 import { KeywordType } from '@v6y/core-logic';
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import { Card, Checkbox } from 'antd';
 import { CheckboxOptionType } from 'antd/es/checkbox/Group';
 import * as React from 'react';
@@ -11,7 +12,6 @@ import {
 import useNavigationAdapter from '../../../infrastructure/adapters/navigation/useNavigationAdapter';
 import VitalityApiConfig from '../../config/VitalityApiConfig';
 import VitalityEmptyView from '../VitalityEmptyView';
-import VitalityLoader from '../VitalityLoader';
 import GetIndicatorListByParams from './getIndicatorListByParams';
 
 interface VitalitySelectableIndicatorsQueryType {
