+++ v6y-apps/front/src/features/apps-stats/components/VitalityAppsStatsView.tsx
@@ -1,9 +1,9 @@
 'use client';
 
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import * as React from 'react';
 import { Suspense } from 'react';
 
-import VitalityLoader from '../../../commons/components/VitalityLoader';
 import VitalitySelectableIndicators from '../../../commons/components/indicators/VitalitySelectableIndicators';
 import VitalityAppList from '../../app-list/components/VitalityAppList';
 import VitalityAppsStatsChart from './VitalityAppsStatsChart';
