+++ v6y-apps/front/src/features/search/components/VitalitySearchView.tsx
@@ -1,7 +1,7 @@
+import VitalityLoader from '@v6y/shared-ui/src/components/VitalityLoader/VitalityLoader';
 import * as React from 'react';
 import { Suspense } from 'react';
 
-import VitalityLoader from '../../../commons/components/VitalityLoader';
 import VitalityAppList from '../../app-list/components/VitalityAppList';
 
 const VitalitySearchView = ({}) => (
