+++ v6y-libs/shared-ui/src/components/VitalityLoader/VitalityLoader.tsx
