+++ v6y-libs/shared-ui/src/index.ts
@@ -1,4 +1,5 @@
+import * as VitalityLoader from './components/VitalityLoader/VitalityLoader.tsx';
 import VitalityText from './components/VitalityText/VitalityText.tsx';
 import VitalityTitle from './components/VitalityTitle/VitalityTitle.tsx';
 
-export { VitalityText, VitalityTitle };
+export { VitalityText, VitalityTitle, VitalityLoader };
